workers: split into separate files and modernise standard Worker methods
